fix(examples/tls-rustls): enable alpn_protocols + cleanup
